app 0.1v:
normal flow:
ask for source files -> new filesongs
from folder temp or another folder
convert to FIleSong, then requests in batches of 10 consecutive requests
then select the correct track from the responses
then in asking in batches of 10 request of a maximum of 50 artists, asks api, add to database(as item, as artist, its genres) and repositories
then in batches of then of maximum 20 albums, ask
 api then stores in database(as item, as album, its genres, its artists in albumArtists-table and repositories
 finally, when the foreign keys are in the database and repositories converts filesong, response, spotifytrack to Track and stores in both database and repositories

ask from database -> files in database
filter by directory, then by response status
(no checked, checked not contained, checked contained, added) and starts the search for artists and albums

it uses a lot of maps and sets to ensure no unnecessary consult to api will be made.

at the initialization of the app, creates every resource, once main starts, waits 10 seconds just in case, then reads database and fills repository, acting as a register to keep the data in memory. currently there's up to 5000 files to add, so an estimate of 5-6 years before it reaches 10000.

in that case the database might be read lazy-mode
but for now its read as a whole and stored all in memory.

next:
1. make the rest of the flows
2. currently JAudioTagger is only used for reading title,album,artist from a file but is intended to also write the data once all the relations from a track are in the database,
that implies a new status for the Response in the sense "tagged"
3. intended to add WEB_images as links in the database from the api, for the future store the images and put them as portraits of the songs, so also in that last state as Local_images